+++ common/src/common/java/info/u_team/music_player/MusicPlayerModReference.java
@@ -1,0 +1,7 @@
+package info.u_team.music_player;
+
+public class MusicPlayerModReference {
+	
+	public static final String MODID = "musicplayer";
+	
+}