+++ fabric/src/main/java/info/u_team/music_player/MusicPlayerMod.java
@@ -1,0 +1,19 @@
+package info.u_team.music_player;
+
+import org.slf4j.Logger;
+
+import com.mojang.logging.LogUtils;
+
+import info.u_team.music_player.init.MusicPlayerClientConstruct;
+import net.fabricmc.api.ClientModInitializer;
+
+public class MusicPlayerMod implements ClientModInitializer {
+	
+	public static final String MODID = MusicPlayerModReference.MODID;
+	public static Logger LOGGER = LogUtils.getLogger();
+	
+	@Override
+	public void onInitializeClient() {
+		MusicPlayerClientConstruct.construct();
+	}
+}