+++ forge/src/main/java/info/u_team/music_player/MusicPlayerMod.java
@@ -1,0 +1,17 @@
+package info.u_team.music_player;
+
+import info.u_team.u_team_core.util.annotation.AnnotationManager;
+import info.u_team.u_team_core.util.verify.JarSignVerifier;
+import net.minecraftforge.fml.common.Mod;
+
+@Mod(MusicPlayerMod.MODID)
+public class MusicPlayerMod {
+	
+	public static final String MODID = MusicPlayerModReference.MODID;
+	
+	public MusicPlayerMod() {
+		JarSignVerifier.checkSigned(MODID);
+		
+		AnnotationManager.callAnnotations(MODID);
+	}
+}