+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [1.19.4-2.5.2.204] - 2023-03-28
+## [1.19.4-2.6.0.205-SNAPSHOT] - 2023-03-29
 ### Changed
- - Port to 1.19.4
+ - Make multi loader project
+ - Release fabric and forge version not together